backend/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyMySQL==1.1.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
